remove failure detector
